+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Recommended Gamevault Server Version: `v14.0.0`
 ### Changes
 
 - Cloud Saves (+)
-- Added Community Themes link to the settings
+- Added retrieval and installation of community themes
+- Added Community Themes Repository link to the settings 
 - Bug fix: Youtube url not loading in gameview
 - Bug fix: Game window sometimes started in the background when launched via gamevault uri
 - Bug fix: Crash on saving user details